siterip presets, search for categories.
each of the 3 numbers is a category, it has to have 1 for on or 0 for
off in the search.
general is number 1, anime is number 2, people is number 3.
if you want to make a search for just anime then search for 010 for
example. 001 for just people or 100 for general.
you can search in any combination.