furaffinity updated and gallery workarounds pushed
cleaned up furaffinity for creator and title tags (it was getting multiple)
been having some 403 errors on this one and i have no idea why. @rachmadaniharyono i would love if you could take a look. the 403 error happens on both the booru workaround and the url downloader. but the test part of the parser window works perfectly.