updated moe parser for sankaku and removed sankaku specific parser.
chan.sankaku and idol.sankaku both uses moebooru as a base and works with that. added examples to moebooru parser to auto match with url classes and removed the unneeded sankaku parser we had. made sure to move over everything.
corrected the time imported to be local instead of GMT which was wrong on all sites by the looks of it.
added missing namespaces, medium, booru, idol, photo_set, rating and studio.
also added support for flash files (with md5 hash).
also fixed the photo_set namespace as that was broken at some point by a typo. wew