derpi booru gallery json parser.
note that it can grab everything but the tags on its own but does also grab the post url and will automatically use that aswell to grab tags.
you might be able to speed it up by removing the extra parsing options for it.
but ill leave it in till we can get a fix on those tags. (it was getting every tag as a single line)