zerochan.net url classes and parsers
url classes automatically adds p=1 for page 1 and s=id to sort the newest. the other url class then matches that and is used by the gallery parser.

testing has been successful.
it wont rip source time for those that doesnt list it in attribute, needs to pull sting or html and clean it up. but i failed to do so when reusing older regexes.