nxtcomicsclub.com url classes and parsers
#61 nxtcomicsclub.com url classes and parsers
grabs all posts from galleries and all urls from posts.
ties it all to either the cover or one of the preview images if cover is offline.
will prob fail if there is no cover or preview. unsure how to fix this. maybe by having a global image like the site logo as the lowest tier.