3dbooru (behoimi.org) removed one url class as both was matching one.
noticed i had some gallery dupes forming an infinite loop of gallery parsing. hopefully this will fix it. will requery the one i had issues with to see if it resolved or not.
the other url class matches this one too so it was always one too many for this one.