**IMPORTANT** yiff.party url association fix!
this will remove an bug/oversight in the last version.
it DID grab the urls in the body of the post but it did so with association and didnt try to download it. (it tried to do both but failed)
users might be missing out on files cause of this error and needs to delete the url from the file its tied to before they can try downloading it again.