added new md5 url class to link with the post parser.
added new md5 url class to link with the post parser.
results from iqdb gave out urls similar to the gallery url class but with the addisonal md5 param which would send the browser to the post url of that md5 value. this parser was a simple duplication of the gallery url class with the addition of requiring the md5 prarm and set to a post url class.