uploaded Suika's Sankaku API parser
This parser will use the native api of the two sites.
It will grab posts 100 at a time from queries.
Posts will have a timelimit from the gallery parse (a day?).
This will download files VERY VERY fast, this might be a risky way to use the site and might get you banned or timed out at a minimum.

After a month or so of testing, it seems pretty stable.

Cause of the time limit, it does not handle resuming of sessions a day or two later to resume the download well at all.
The gallery parser need to re-parse for new post urls.