Fix alternative known urls not being added + index url recognition
Overlooked that if you don't set the "know urls" on some types of posts, they will not be added to the links, even if they are present in the parser.
Per that change 2 post urls were changed from gallery_url to post_url... it was a troubling time back then, I swear.
And the index pages of both chan and beta are being recognized.