Add page tag to books/pools in sankaku pool parser
Yeah, about that. The pools now known as books have a sequnce id in the api, so that will be used to tag images with pages. Depends if one really should use it, but since books are getting bigger on sankaku and in more automated ways, it's far outweights the contra argument of not adding pages to pools because they would be created by users.

TL;DR download using proper tags and only use pools/books for content that is definitely in the right order