fix kemono.party body text url parsing
When an artist posted a link where the link text itself was the URL the link leads to, it would generate two URLs to parse, one of which had HTML garbage in it that made it unusable (i.e. you'd get https://link.to/stuff and https://link.to/stuff</a></p>). This changes the parser to select the URL parsed from the <a> tag's href text so it will only generate one URL per link.