fix kemono.party url parsing
When an artist posted a link where the link text itself was the URL the link leads to, it would generate two URLs to follow, one of which had HTML garbage in it that made it unusable (i.e. you'd get both https://link.to/stuff and https://link.to/stuff</a></p>). This changes the parser to specifically select the URL parsed from the tag's href text so it will only generate one URL per link.