further updates to kemono.party parser
This fixes an issue where older files which are given with absolute paths in the JSON would have malformed URLs due to `https://data.kemono.party` being appended to the front (resulting in URLs of the form `https://data.kemono.partyhttps://kemono.party/path/to/file`). This is done with a regex that checks for this type of malformed URL and corrects it by removing the `https://kemono.party` part.

This also improves URL parsing from body text by cleaning up extraneous text that can end up in front of the URL, as was the case in the example URL provided with the parser previously. It would find a line with a dropbox URL on it, but it would have a bunch of junk in front of it which resulted in passing a malformed URL to the gallery downloader.