New parsers for baraag and pawoo
The current parser in the repo just grabs the bare URLs from the artist's media feed and associates the username with them. These new parsers grab the post URLs and associate them and their post times with the downloaded media, and also grab hashtags from the posts (which on these websites tend to be content descriptors and not total garbage like Twitter hashtags).