More features, change parser names to "mastodon"
Added parsing for the displayed names that users can change in addition to the fixed username, and also added the parsing for source instance from the other mastodon parser. Changed the parser names to "mastodon" so-and-so parser to reflect that they are generic parsers for any mastodon instance as long as you have the appropriate URL classes.