updated kemono.party!
Updated two things in this update. I removed the subdomain jank we had in the content parsers mentioned here: https://github.com/CuddleBear92/Hydrus-Presets-and-Scripts/issues/105#issuecomment-903954370
Also added an nGUG for all the creator ID's allowing you to easily use a single gallery query for any of the supported sites... ofc this will generate errors as it fails to find ids for a specific site.
Also it might cause another creator to be downloaded, please try with caution.