Updated kemono.party/coomer.party for new API
New URL classes and fixes to the parser to support the new API endpoint. Old API URL classes are redirected to the new endpoints to avoid 404s for people who have old URLs still waiting to be processed in downloaders/subscriptions. I've also rolled in the post description note parsing and parsing of plaintext URLs in the description (e.g. when an artist posts a bare link to a Google drive or something) from my personal modifications. The favorites parser is also fixed and now supports downloading favorited posts in addition to favorited artists.