updated kemono & coomer
removed un-needed api url classes and added redirections to the .party domain incase the site stops redirecting on its own in the future.

did some changes to the login script but still doesn't work, unsure if its the ddg1 or session cookie that considers a login.... and the session cookie is not getting added at all for whatever reason.