updated zerochan.net
fixed up the base gallery url class to use the param system instead of what it was doing (it was appending just a simple string). Now the url class will always use page 1 and the id sort order by default but also do pagenation navigation on its own as a backup from the parser.
the gallery parser also had weird priority values so they are equal. debating removing the pagenation url parser in the gallery parser as its technically not needed but i left it in for now.