Update fantia parser to handle CSRF tokens
Shoutouts to gallery-dl and Cultered-Downloader-Logic for showing the new
authentication requirements for fantia's API. The way to do this in a Hydrus
parser is a bit weird: we can no longer use API redirects in our URL classes,
but still have to have URL classes and parsers for everything, and the post
URL parser has to basically just generate another post URL (an API one) along
with the CSRF header required to auth with the API. But, it works. I even
restored the lost behavior from when this downloader underwent a revision
that dropped the ability to download thumbnails. If you don't like the thumbs,
just delete the sub-parser for thumbs inside the sub-parser for the content.